GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   - display.py
   - config.py
   - main.py
   - aco.py
   - weather_data.nc
   - contrail_grid.nc

Your previous branch was: refs/heads/main

The sha for that commit was: 95b272815144dba493be724f8a086a99382011a9

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch